+++ plugins/cookie_manager/lib/dio_cookie_manager.dart
@@ -1,3 +1,4 @@
 library dio_cookie_manager;
 
 export 'src/cookie_mgr.dart';
+export 'src/exception.dart';