+++ plugins/cookie_manager/lib/src/cookie_mgr.dart
@@ -4,6 +4,8 @@ import 'dart:io';
 import 'package:cookie_jar/cookie_jar.dart';
 import 'package:dio/dio.dart';
 
+import 'exception.dart';
+
 const _kIsWeb = bool.hasEnvironment('dart.library.js_util')
     ? bool.fromEnvironment('dart.library.js_util')
     : identical(0, 0.0);
@@ -25,6 +27,11 @@ class CookieManager extends Interceptor {
     this.cookieJar,
   ) : assert(!_kIsWeb, "Don't use the manager in Web environments.");
 
+  /// The cookie jar used to load and save cookies.
+  ///
+  /// See also:
+  /// * [CookieJar]
+  /// * [PersistCookieJar]
   final CookieJar cookieJar;
 
   /// Merge cookies into a Cookie string.
@@ -46,73 +53,116 @@ class CookieManager extends Interceptor {
   }
 
   @override
-  void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
-    cookieJar.loadForRequest(options.uri).then((cookies) {
-      final previousCookies =
-          options.headers[HttpHeaders.cookieHeader] as String?;
-      final newCookies = getCookies([
-        ...?previousCookies
-            ?.split(';')
-            .where((e) => e.isNotEmpty)
-            .map((c) => Cookie.fromSetCookieValue(c)),
-        ...cookies,
-      ]);
+  Future<void> onRequest(
+    RequestOptions options,
+    RequestInterceptorHandler handler,
+  ) async {
+    try {
+      final cookies = await loadCookies(options);
       options.headers[HttpHeaders.cookieHeader] =
-          newCookies.isNotEmpty ? newCookies : null;
+          cookies.isNotEmpty ? cookies : null;
       handler.next(options);
-    }).catchError((dynamic e, StackTrace s) {
-      final error = DioException(
+    } catch (e, s) {
+      final exception = DioException(
         requestOptions: options,
-        error: e,
-        stackTrace: s,
+        type: DioExceptionType.unknown,
+        error: CookieManagerLoadException(error: e, stackTrace: s),
+        message: 'Failed to load cookies for the request.',
       );
-      handler.reject(error, true);
-    });
-  }
-
-  @override
-  void onResponse(Response response, ResponseInterceptorHandler handler) {
-    _saveCookies(response).then((_) => handler.next(response)).catchError(
-      (dynamic e, StackTrace s) {
-        final error = DioException(
-          requestOptions: response.requestOptions,
-          error: e,
-          stackTrace: s,
-        );
-        handler.reject(error, true);
-      },
-    );
-  }
-
-  @override
-  void onError(DioException err, ErrorInterceptorHandler handler) {
-    if (err.response != null) {
-      _saveCookies(err.response!).then((_) => handler.next(err)).catchError(
-        (dynamic e, StackTrace s) {
-          final error = DioException(
-            requestOptions: err.response!.requestOptions,
-            error: e,
-            stackTrace: s,
-          );
-          handler.next(error);
-        },
-      );
-    } else {
-      handler.next(err);
+      handler.reject(exception, true);
     }
   }
 
-  Future<void> _saveCookies(Response response) async {
+  @override
+  Future<void> onResponse(
+    Response response,
+    ResponseInterceptorHandler handler,
+  ) async {
+    try {
+      await saveCookies(response);
+      handler.next(response);
+    } catch (e, s) {
+      final exception = DioException(
+        requestOptions: response.requestOptions,
+        response: response,
+        type: DioExceptionType.unknown,
+        error: CookieManagerSaveException(
+          response: response,
+          error: e,
+          stackTrace: s,
+          dioException: null,
+        ),
+        stackTrace: s,
+      );
+      handler.reject(exception, true);
+    }
+  }
+
+  @override
+  Future<void> onError(
+    DioException err,
+    ErrorInterceptorHandler handler,
+  ) async {
+    final response = err.response;
+    if (response == null) {
+      handler.next(err);
+      return;
+    }
+
+    try {
+      await saveCookies(response);
+      handler.next(err);
+    } catch (e, s) {
+      final exception = DioException(
+        requestOptions: response.requestOptions,
+        response: response,
+        type: DioExceptionType.unknown,
+        error: CookieManagerSaveException(
+          response: response,
+          error: e,
+          stackTrace: s,
+          dioException: err,
+        ),
+        stackTrace: s,
+      );
+      handler.next(exception);
+    }
+  }
+
+  /// Load cookies in cookie string for the request.
+  Future<String> loadCookies(RequestOptions options) async {
+    final savedCookies = await cookieJar.loadForRequest(options.uri);
+    final previousCookies =
+        options.headers[HttpHeaders.cookieHeader] as String?;
+    final cookies = getCookies([
+      ...?previousCookies
+          ?.split(';')
+          .where((e) => e.isNotEmpty)
+          .map((c) => Cookie.fromSetCookieValue(c)),
+      ...savedCookies,
+    ]);
+    return cookies;
+  }
+
+  /// Save cookies from the response including redirected requests.
+  Future<void> saveCookies(Response response) async {
     final setCookies = response.headers[HttpHeaders.setCookieHeader];
     if (setCookies == null || setCookies.isEmpty) {
       return;
     }
+
     final List<Cookie> cookies = setCookies
         .map((str) => str.split(_setCookieReg))
         .expand((cookie) => cookie)
         .where((cookie) => cookie.isNotEmpty)
         .map((str) => Cookie.fromSetCookieValue(str))
         .toList();
+    // Saving cookies for the original site.
+    // Spec: https://www.rfc-editor.org/rfc/rfc7231#section-7.1.2.
+    final originalUri = response.requestOptions.uri;
+    final realUri = originalUri.resolveUri(response.realUri);
+    await cookieJar.saveFromResponse(realUri, cookies);
+
     // Handle `Set-Cookie` when `followRedirects` is false
     // and the response returns a redirect status code.
     final statusCode = response.statusCode ?? 0;
@@ -121,13 +171,8 @@ class CookieManager extends Interceptor {
     // We don't want to explicitly consider recursive redirections
     // cookie handling here, because when `followRedirects` is set to false,
     // users will be available to handle cookies themselves.
-    final isRedirectRequest = statusCode >= 300 && statusCode < 400;
-    // Saving cookies for the original site.
-    // Spec: https://www.rfc-editor.org/rfc/rfc7231#section-7.1.2.
-    final originalUri = response.requestOptions.uri;
-    final realUri = originalUri.resolveUri(response.realUri);
-    await cookieJar.saveFromResponse(realUri, cookies);
-    if (isRedirectRequest && locations.isNotEmpty) {
+    final redirected = statusCode >= 300 && statusCode < 400;
+    if (redirected && locations.isNotEmpty) {
       final originalUri = response.realUri;
       await Future.wait(
         locations.map(